Refactor to use api gateway validation.
