--- layers/ajv/nodejs/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "swb-tools-layers-ajv",
-  "version": "1.0.0",
-  "description": "Ajv JSON schema validator for lambdas",
-  "dependencies": {
-    "ajv": "^8.11.0",
-    "ajv-formats": "^2.1.1",
-    "ajv-errors": "^3.0.0"
-  }
-}
-
