+++ registration/api/registration.schema.json
@@ -1,0 +1,28 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "title": "Registration",
+  "description": "A registering user's information",
+  "type": "object",
+  "additionalProperties": false,
+  "properties": {
+    "firstName": {
+      "type": "string",
+      "minLength": 1,
+      "maxLength": 100,
+      "pattern": "^[A-Za-z0-9 .-]+$"
+    },
+    "lastName": {
+      "type": "string",
+      "minLength": 1,
+      "maxLength": 100,
+      "pattern": "^[A-Za-z0-9 .-]+$"
+    },
+    "email": {
+      "type": "string",
+      "minLength": 5,
+      "maxLength": 254,
+      "format": "email"
+    }
+  },
+  "required": [ "firstName", "lastName", "email" ]
+}
