+++ registration/api/serverless.yml
@@ -5,32 +5,38 @@ frameworkVersion: "3"
 provider:
   name: "aws"
   runtime: "nodejs14.x"
+  apiGateway:
+      request:
+        schemas:
+          register-user:
+            name: RegisterUser
+            schema: ${file(./registration.schema.json)}
+            description: "A validation model for registering users."
 
 package:
   patterns:
     - "!node_modules/**"
-    - "!package.json"
+    - "!__snapshots__/*"
     - "!template.js"
+    - "!*.test.js"
+    - "!package.json"
+    - "!*.schema.json"
     - "!pnpm-lock.yaml"
-    - "!test/*"
-    - "!test/__snapshots__/*"
 
 custom:
   config: "${file(../../configs/${sls:stage}.yml)}"
   info: "${file(./package.json)}"
-  layer-avj: "${file(../../layers/ajv/nodejs/package.json)}"
   layer-lodash: "${file(../../layers/lodash/nodejs/package.json)}"
   layer-uuid: "${file(../../layers/uuid/nodejs/package.json)}"
   userTableArn: "arn:aws:dynamodb:${aws:region}:${aws:accountId}:table/${self:custom.config.userTableName}"
 
 functions:
   AccountRegistrationNotification:
-    handler: "src/handler.registerUser"
+    handler: "handler.registerUser"
     name: "${self:service}-${sls:stage}"
     description: "${self:custom.info.description}"
     layers:
       - "arn:aws:lambda:${aws:region}:${aws:accountId}:layer:${layerName(${self:custom.layer-lodash})}-${sls:stage}:1"
-      - "arn:aws:lambda:${aws:region}:${aws:accountId}:layer:${layerName(${self:custom.layer-avj})}-${sls:stage}:1"
       - "arn:aws:lambda:${aws:region}:${aws:accountId}:layer:${layerName(${self:custom.layer-uuid})}-${sls:stage}:1"
     iamRoleStatementsName: "${self:service}-${sls:stage}-lambda"
     iamRoleStatements:
@@ -42,6 +48,9 @@ functions:
       - http:
           path: api/register
           method: post
+          request:
+            schemas:
+              application/json: register-user
           # cors:
           #   origins:
           #     - "http://localhost:3000"
